services/registry/requirements.txt
nameko
nameko-redis
nameko-sqlalchemy
PyMySQL
sqlalchemy_utils
